make components directory
