+++ tic-tac-toe-app/src/components/Board.tsx
@@ -1,0 +1,36 @@
+import Square from './Square'
+import { FC } from 'react';
+
+type BoardProps = {
+  squares:  Array<string | null>
+  onClick: (i: number) => void
+}
+const Board: FC<BoardProps> = ({ squares, onClick }) => {
+  const renderSquare = (i : number) => {
+    return <Square 
+    value={squares[i]}
+    onClick={() => onClick(i)}
+    />;
+  }
+  return (
+    <div>
+      <div className="board-row">
+        {renderSquare(0)}
+        {renderSquare(1)}
+        {renderSquare(2)}
+      </div>
+      <div className="board-row">
+        {renderSquare(3)}
+        {renderSquare(4)}
+        {renderSquare(5)}
+      </div>
+      <div className="board-row">
+        {renderSquare(6)}
+        {renderSquare(7)}
+        {renderSquare(8)}
+      </div>
+    </div>
+  );
+}
+
+export default Board
